+++ frontend_tests/node_tests/user_events.js
@@ -1,6 +1,7 @@
 set_global('$', global.make_zjquery());
 
 zrequire('people');
+const settings_config = zrequire('settings_config');
 zrequire('user_events');
 
 set_global('activity', {
@@ -75,14 +76,17 @@ run_test('updates', () => {
     };
     people.add_active_user(isaac);
 
-    user_events.update_person({user_id: isaac.user_id, is_guest: true});
+    user_events.update_person({user_id: isaac.user_id,
+                               role: settings_config.user_role_values.guest.code});
     person = people.get_by_email(isaac.email);
     assert(person.is_guest);
-    user_events.update_person({user_id: isaac.user_id, is_guest: false});
+    user_events.update_person({user_id: isaac.user_id,
+                               role: settings_config.user_role_values.member.code});
     person = people.get_by_email(isaac.email);
     assert(!person.is_guest);
 
-    user_events.update_person({user_id: isaac.user_id, is_admin: true});
+    user_events.update_person({user_id: isaac.user_id,
+                               role: settings_config.user_role_values.admin.code});
     person = people.get_by_email(isaac.email);
     assert.equal(person.full_name, 'Isaac Newton');
     assert.equal(person.is_admin, true);
@@ -101,7 +105,8 @@ run_test('updates', () => {
     assert.equal(user_id, isaac.user_id);
     assert.equal(full_name, 'Sir Isaac');
 
-    user_events.update_person({user_id: me.user_id, is_admin: false});
+    user_events.update_person({user_id: me.user_id,
+                               role: settings_config.user_role_values.member.code});
     assert(!global.page_params.is_admin);
 
     user_events.update_person({user_id: me.user_id, full_name: 'Me V2'});